fix(dashboard): lazy-load recharts to keep main bundle under 2 MiB
The web build failed with a vite-plugin-pwa Workbox precache error because
the main `index-*.js` chunk crossed 2 MiB after recharts was added. Splitting
the chart into its own module loaded via React.lazy moves recharts to a
separate `pnl-graph-chart-*.js` chunk that is only fetched when the widget
renders.

Build output after fix:
- dist/assets/pnl-graph-chart-*.js  342 kB
- dist/assets/index-*.js          1,812 kB  (under the 2 MiB precache limit)
